move collection stores
